Push proposal transaction #644

+++ libraries/chain/include/golos/chain/operation_notification.hpp
@@ -18,7 +18,7 @@ namespace golos {
             uint32_t block = 0;
             uint32_t trx_in_block = 0;
             uint16_t op_in_trx = 0;
-            uint64_t virtual_op = 0;
+            uint32_t virtual_op = 0;
             const operation &op;
         };
